Backend/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.1
djangorestframework>=3.15
python-dotenv>=1.0
django-cors-headers>=4.0
pytest-django>=4.5